pyproject.toml
[tool.poetry]
name = "ranobelib-parser"
version = "1.0"
description = ""
authors = ["Midnull"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
iso-639 = {git = "https://github.com/noumar/iso639.git", tag = "0.4.5"}
fb2 = "^0.1.7"
setuptools = "^71.1.0"
beautifulsoup4 = "^4.12.3"
cloudscraper = "^1.2.71"
pillow = "^10.4.0"
ebooklib = "^0.18"
textual = "^0.74.0"
textual-fspicker = "^0.0.11"
pyperclip = "^1.9.0"
pyjwt = "^2.9.0"

[tool.ruff]

line-length = 120
target-version = "py312"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
